Split VPN copy used by main app and shared with VPN targets
