+++ LocalPackages/NetworkProtectionMac/Sources/NetworkProtectionUI/Resources/Localizable.xcstrings
@@ -3475,7 +3475,7 @@
     },
     "network.protection.vpn.location.nearest" : {
       "comment" : "Description of the location type in the VPN status view",
-      "extractionState" : "extracted_with_value",
+      "extractionState" : "stale",
       "localizations" : {
         "de" : {
           "stringUnit" : {
